Move mine back to root
